[MER-3180] Implement Practice Activities pagination. Fix related tests.
