+++ test/oli_web/live/delivery/instructor_dashboard/overview/practice_activities_tab_test.exs
@@ -2034,7 +2034,7 @@ defmodule OliWeb.Delivery.InstructorDashboard.PracticeActivitiesTabTest do
 
       # Change page size from default (20) to 2
       view
-      |> element("#header_paging_page_size_form")
+      |> element("#footer_paging_page_size_form")
       |> render_change(%{limit: "2"})
 
       [a0, a1] = table_as_list_of_maps(view)
@@ -2065,7 +2065,7 @@ defmodule OliWeb.Delivery.InstructorDashboard.PracticeActivitiesTabTest do
 
       # Change page size from 2 to 1
       view
-      |> element("#header_paging_page_size_form")
+      |> element("#footer_paging_page_size_form")
       |> render_change(%{limit: "1"})
 
       [a2] = table_as_list_of_maps(view)
